requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
crispy-tailwind==1.0.3
cryptography==43.0.3
decouple==0.0.7
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.5
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
swagger-spec-validator==3.0.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
